feat(enriched-readings-topic): publish enriched readings to the dionysus/enrichedReadings topic when they are ingested

+++ common/src/main/kotlin/com/dionysus/common/domain/Event.kt
@@ -1,3 +1,3 @@
 package com.dionysus.common.domain
 
-class Event(val valveId: Int, val value: Double) : Message
+class Event(val valveId: Int, val value: Int) : Message
